Fix #1694, Support DVR 2GB+ MP4 file. 3.0.155

+++ trunk/src/kernel/srs_kernel_buffer.hpp
@@ -62,19 +62,12 @@ public:
     ISrsCodec();
     virtual ~ISrsCodec();
 public:
-    /**
-     * get the number of bytes to code to.
-     */
-    // TODO: FIXME: change to uint64_t.
-    virtual int nb_bytes() = 0;
-    /**
-     * encode object to bytes in SrsBuffer.
-     */
+    // Get the number of bytes to code to.
+    virtual uint64_t nb_bytes() = 0;
+    // Encode object to bytes in SrsBuffer.
     virtual srs_error_t encode(SrsBuffer* buf) = 0;
 public:
-    /**
-     * decode object from bytes in SrsBuffer.
-     */
+    // Decode object from bytes in SrsBuffer.
     virtual srs_error_t decode(SrsBuffer* buf) = 0;
 };
